==> R_scripts/2_descriptive_statistics.Rmd <==
---
title: "Descriptive statistics"
author: "Gaia Peresini"
date: '`r Sys.Date()`'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(include = FALSE)
```

```{r}
# Libraries
library(tidyverse)
library(data.table)
```


```{r}
# Load data
lisa_dt <- readRDS(file = "//.../Original_dt/lisa_dt.rds")
bkg_dt <- readRDS(file = "//.../Original_dt/bkg_dt.rds")
hushall_dt <- readRDS(file = "//.../Original_dt/hushall_dt.rds")
geo_dt <- readRDS(file = "//.../Original_dt/geo_dt.rds")

deso_stats <- readRDS(file = "//.../deso_stats.rds")
new_building_vacancies2 <- readRDS(file = "//.../new_building_vacancies2.rds")
kommun_vec <- readRDS(file = "//.../kommun_vec.rds")
```

```{r}
# Population
lisa_dt %>%
  filter(Kommun %in% kommun_vec) %>%
  group_by(Ar) %>%
  count()

# Income
lisa_dt %>%
  mutate(Stkm = ifelse(Kommun %in% kommun_vec, 1, 0)) %>% # 1: Stockholm county, 0 rest of Sweden
  group_by(Ar) %>% #Stkm
  summarize(Mean = mean(income, na.rm = T),
            SD = sd(income, na.rm = T),
            "NA" = sum(is.na(income)))
  
# Country of birth
lisa_dt %>%
  left_join(bkg_dt, by="PersonalNr") %>%
  filter(Kommun %in% kommun_vec) %>%
  group_by(Ar, swe_born) %>%
  summarize(n = n()) %>%
  mutate(freq = n / sum(n)*100)
  # count(Ar, swe_born) %>%
  # mutate(freq = prop.table(n), .by=Ar) 
  
# Sex
lisa_dt %>%
  left_join(bkg_dt, by="PersonalNr") %>%
  filter(Kommun %in% kommun_vec) %>%
  group_by(Ar, sex) %>%
  summarize(n = n()) %>%
  mutate(freq = n / sum(n)*100)

# Education
lisa_dt %>%
  filter(Kommun %in% kommun_vec) %>%
  group_by(Ar, ma_above) %>%
  distinct(PersonalNr) %>%
  summarize(n = n()) %>%
  mutate(freq = n / sum(n)*100)

# New buildings
new_building_vacancies2 %>%
  mutate(Year = year_of_building_creation) %>%
  group_by(Year) %>%
  count()

# DeSO
geo_dt %>%
  filter(Kommun %in% kommun_vec) %>%
  group_by(Ar) %>%
  distinct(BostDESO) %>%
  count()
```
